removed hardcoded exp data
